requirements.txt
vosk
scipy
numpy
pyttsx3
python-dotenv
openai-agents[voice]